Move proofs endpoints to new namespace
