+++ packages/api/src/beacon/client/index.ts
@@ -9,6 +9,7 @@ import * as events from "./events.js";
 import * as lightclient from "./lightclient.js";
 import * as lodestar from "./lodestar.js";
 import * as node from "./node.js";
+import * as proofs from "./proofs.js";
 import * as validator from "./validator.js";
 
 type ClientModules = HttpClientModules & {
@@ -31,6 +32,7 @@ export function getClient(opts: HttpClientOptions, modules: ClientModules): Api 
     lightclient: lightclient.getClient(config, httpClient),
     lodestar: lodestar.getClient(config, httpClient),
     node: node.getClient(config, httpClient),
+    proofs: proofs.getClient(config, httpClient),
     validator: validator.getClient(config, httpClient),
   };
 }
