+++ packages/api/src/beacon/client/lightclient.ts
@@ -1,7 +1,6 @@
 import {IChainForkConfig} from "@lodestar/config";
-import {deserializeProof} from "@chainsafe/persistent-merkle-tree";
 import {Api, ReqTypes, routesData, getReqSerializers, getReturnTypes} from "../routes/lightclient.js";
-import {IHttpClient, getFetchOptsSerializers, generateGenericJsonClient} from "../../utils/client/index.js";
+import {IHttpClient, generateGenericJsonClient} from "../../utils/client/index.js";
 
 /**
  * REST HTTP client for lightclient routes
@@ -9,19 +8,6 @@ import {IHttpClient, getFetchOptsSerializers, generateGenericJsonClient} from ".
 export function getClient(_config: IChainForkConfig, httpClient: IHttpClient): Api {
   const reqSerializers = getReqSerializers();
   const returnTypes = getReturnTypes();
-
-  // Some routes return JSON, use a client auto-generator
-  const client = generateGenericJsonClient<Api, ReqTypes>(routesData, reqSerializers, returnTypes, httpClient);
-  // For `getStateProof()` generate request serializer
-  const fetchOptsSerializers = getFetchOptsSerializers<Api, ReqTypes>(routesData, reqSerializers);
-
-  return {
-    ...client,
-
-    async getStateProof(stateId, paths) {
-      const buffer = await httpClient.arrayBuffer(fetchOptsSerializers.getStateProof(stateId, paths));
-      const proof = deserializeProof(new Uint8Array(buffer));
-      return {data: proof};
-    },
-  };
+  // All routes return JSON, use a client auto-generator
+  return generateGenericJsonClient<Api, ReqTypes>(routesData, reqSerializers, returnTypes, httpClient);
 }
