+++ packages/api/src/beacon/index.ts
@@ -16,6 +16,7 @@ const allNamespacesObj: {[K in keyof Api]: true} = {
   lightclient: true,
   lodestar: true,
   node: true,
+  proofs: true,
   validator: true,
 };
 export const allNamespaces = Object.keys(allNamespacesObj) as ApiNamespace[];
