+++ packages/api/src/beacon/routes/lightclient.ts
@@ -1,18 +1,14 @@
-import {JsonPath} from "@chainsafe/ssz";
-import {Proof} from "@chainsafe/persistent-merkle-tree";
 import {altair, phase0, ssz, SyncPeriod} from "@lodestar/types";
 import {
   ArrayOf,
   ReturnTypes,
   RoutesData,
   Schema,
-  sameType,
   ContainerData,
   ReqSerializers,
   reqEmpty,
   ReqEmpty,
 } from "../../utils/index.js";
-import {queryParseProofPathsArr, querySerializeProofPathsArr} from "../../utils/serdes.js";
 
 // See /packages/api/src/routes/index.ts for reasoning and instructions to add new routes
 
@@ -24,11 +20,6 @@ export type LightClientBootstrap = {
 };
 
 export type Api = {
-  /**
-   * Returns a multiproof of `jsonPaths` at the requested `stateId`.
-   * The requested `stateId` may not be available. Regular nodes only keep recent states in memory.
-   */
-  getStateProof(stateId: string, jsonPaths: JsonPath[]): Promise<{data: Proof}>;
   /**
    * Returns an array of best updates given a `startPeriod` and `count` number of sync committee period to return.
    * Best is defined by (in order of priority):
@@ -59,7 +50,6 @@ export type Api = {
  * Define javascript values for each route
  */
 export const routesData: RoutesData<Api> = {
-  getStateProof: {url: "/eth/v1/beacon/light_client/proof/{state_id}", method: "GET"},
   getUpdates: {url: "/eth/v1/beacon/light_client/updates", method: "GET"},
   getOptimisticUpdate: {url: "/eth/v1/beacon/light_client/optimistic_update", method: "GET"},
   getFinalityUpdate: {url: "/eth/v1/beacon/light_client/finality_update", method: "GET"},
@@ -69,7 +59,6 @@ export const routesData: RoutesData<Api> = {
 
 /* eslint-disable @typescript-eslint/naming-convention */
 export type ReqTypes = {
-  getStateProof: {params: {state_id: string}; query: {paths: string[]}};
   getUpdates: {query: {start_period: number; count: number}};
   getOptimisticUpdate: ReqEmpty;
   getFinalityUpdate: ReqEmpty;
@@ -79,12 +68,6 @@ export type ReqTypes = {
 
 export function getReqSerializers(): ReqSerializers<Api, ReqTypes> {
   return {
-    getStateProof: {
-      writeReq: (state_id, paths) => ({params: {state_id}, query: {paths: querySerializeProofPathsArr(paths)}}),
-      parseReq: ({params, query}) => [params.state_id, queryParseProofPathsArr(query.paths)],
-      schema: {params: {state_id: Schema.StringRequired}, body: Schema.AnyArray},
-    },
-
     getUpdates: {
       writeReq: (start_period, count) => ({query: {start_period, count}}),
       parseReq: ({query}) => [query.start_period, query.count],
@@ -109,8 +92,6 @@ export function getReqSerializers(): ReqSerializers<Api, ReqTypes> {
 
 export function getReturnTypes(): ReturnTypes<Api> {
   return {
-    // Just sent the proof JSON as-is
-    getStateProof: sameType(),
     getUpdates: ContainerData(ArrayOf(ssz.altair.LightClientUpdate)),
     getOptimisticUpdate: ContainerData(ssz.altair.LightClientOptimisticUpdate),
     getFinalityUpdate: ContainerData(ssz.altair.LightClientFinalityUpdate),
