+++ packages/api/src/beacon/server/index.ts
@@ -9,6 +9,7 @@ import * as events from "./events.js";
 import * as lightclient from "./lightclient.js";
 import * as lodestar from "./lodestar.js";
 import * as node from "./node.js";
+import * as proofs from "./proofs.js";
 import * as validator from "./validator.js";
 
 // Re-export for convenience
@@ -36,6 +37,7 @@ export function registerRoutes(
     lightclient: () => lightclient.getRoutes(config, api.lightclient),
     lodestar: () => lodestar.getRoutes(config, api.lodestar),
     node: () => node.getRoutes(config, api.node),
+    proofs: () => proofs.getRoutes(config, api.proofs),
     validator: () => validator.getRoutes(config, api.validator),
   };
