+++ packages/api/test/unit/beacon/testData/lightclient.ts
@@ -1,5 +1,4 @@
 import {ssz} from "@lodestar/types";
-import {ProofType} from "@chainsafe/persistent-merkle-tree";
 import {toHexString} from "@chainsafe/ssz";
 import {Api} from "../../../../src/beacon/routes/lightclient.js";
 import {GenericServerTestCases} from "../../../utils/genericServerTest.js";
@@ -12,30 +11,6 @@ const header = ssz.phase0.BeaconBlockHeader.defaultValue();
 const signatureSlot = ssz.Slot.defaultValue();
 
 export const testData: GenericServerTestCases<Api> = {
-  getStateProof: {
-    args: [
-      "head",
-      [
-        // ["validator", 0, "balance"],
-        ["finalized_checkpoint", 0, "root", 12000],
-      ],
-    ],
-    res: {
-      data: {
-        type: ProofType.treeOffset,
-        offsets: [1, 2, 3],
-        leaves: [root, root, root, root],
-      },
-    },
-    /* eslint-disable quotes */
-    query: {
-      paths: [
-        // '["validator",0,"balance"]',
-        '["finalized_checkpoint",0,"root",12000]',
-      ],
-    },
-    /* eslint-enable quotes */
-  },
   getUpdates: {
     args: [1, 2],
     res: {data: [lightClientUpdate]},
