+++ packages/beacon-node/src/api/impl/api.ts
@@ -8,6 +8,7 @@ import {getEventsApi} from "./events/index.js";
 import {getLightclientApi} from "./lightclient/index.js";
 import {getLodestarApi} from "./lodestar/index.js";
 import {getNodeApi} from "./node/index.js";
+import {getProofsApi} from "./proofs/index.js";
 import {getValidatorApi} from "./validator/index.js";
 
 export function getApi(opts: IApiOptions, modules: ApiModules): Api {
@@ -16,9 +17,10 @@ export function getApi(opts: IApiOptions, modules: ApiModules): Api {
     config: getConfigApi(modules),
     debug: getDebugApi(modules),
     events: getEventsApi(modules),
-    lightclient: getLightclientApi(opts, modules),
+    lightclient: getLightclientApi(modules),
     lodestar: getLodestarApi(modules),
     node: getNodeApi(opts, modules),
+    proofs: getProofsApi(opts, modules),
     validator: getValidatorApi(modules),
   };
 }
