+++ packages/beacon-node/src/api/impl/lightclient/index.ts
@@ -1,48 +1,13 @@
 import {routes} from "@lodestar/api";
 import {fromHexString} from "@chainsafe/ssz";
-import {ProofType, Tree} from "@chainsafe/persistent-merkle-tree";
 import {SyncPeriod} from "@lodestar/types";
 import {MAX_REQUEST_LIGHT_CLIENT_UPDATES, MAX_REQUEST_LIGHT_CLIENT_COMMITTEE_HASHES} from "@lodestar/params";
 import {ApiModules} from "../types.js";
-import {resolveStateId} from "../beacon/state/utils.js";
-import {IApiOptions} from "../../options.js";
 
 // TODO: Import from lightclient/server package
 
-export function getLightclientApi(
-  opts: IApiOptions,
-  {chain, config, db}: Pick<ApiModules, "chain" | "config" | "db">
-): routes.lightclient.Api {
-  // It's currently possible to request gigantic proofs (eg: a proof of the entire beacon state)
-  // We want some some sort of resistance against this DoS vector.
-  const maxGindicesInProof = opts.maxGindicesInProof ?? 512;
-
+export function getLightclientApi({chain}: Pick<ApiModules, "chain">): routes.lightclient.Api {
   return {
-    async getStateProof(stateId, jsonPaths) {
-      const state = await resolveStateId(config, chain, db, stateId);
-
-      // Commit any changes before computing the state root. In normal cases the state should have no changes here
-      state.commit();
-      const stateNode = state.node;
-      const tree = new Tree(stateNode);
-
-      const gindexes = state.type.tree_createProofGindexes(stateNode, jsonPaths);
-      // TODO: Is it necessary to de-duplicate?
-      //       It's not a problem if we overcount gindexes
-      const gindicesSet = new Set(gindexes);
-
-      if (gindicesSet.size > maxGindicesInProof) {
-        throw new Error("Requested proof is too large.");
-      }
-
-      return {
-        data: tree.getProof({
-          type: ProofType.treeOffset,
-          gindices: Array.from(gindicesSet),
-        }),
-      };
-    },
-
     async getUpdates(startPeriod: SyncPeriod, count: number) {
       const maxAllowedCount = Math.min(MAX_REQUEST_LIGHT_CLIENT_UPDATES, count);
       const periods = Array.from({length: maxAllowedCount}, (_ignored, i) => i + startPeriod);
