+++ packages/light-client/test/unit/sync.node.test.ts
@@ -8,7 +8,7 @@ import {chainConfig as chainConfigDef} from "@lodestar/config/default";
 import {createIBeaconConfig, IChainConfig} from "@lodestar/config";
 import {toHexString} from "@chainsafe/ssz";
 import {Lightclient, LightclientEvent} from "../../src/index.js";
-import {LightclientServerApiMock} from "../mocks/LightclientServerApiMock.js";
+import {LightclientServerApiMock, ProofsServerApiMock} from "../mocks/LightclientServerApiMock.js";
 import {EventsServerApiMock} from "../mocks/EventsServerApiMock.js";
 import {
   computeLightclientUpdate,
@@ -58,11 +58,13 @@ describe("sync", () => {
     // Create server impl mock backed
     const lightclientServerApi = new LightclientServerApiMock();
     const eventsServerApi = new EventsServerApiMock();
+    const proofsServerApi = new ProofsServerApiMock();
     // Start server
     const opts: ServerOpts = {host: "127.0.0.1", port: 15000};
     await startServer(opts, config, ({
       lightclient: lightclientServerApi,
       events: eventsServerApi,
+      proofs: proofsServerApi,
     } as Partial<Api>) as Api);
 
     // Populate initial snapshot
@@ -130,7 +132,7 @@ describe("sync", () => {
 
         // Provide the state to the lightclient server impl. Only the last one to test proof fetching
         if (slot === targetSlot) {
-          lightclientServerApi.states.set(toHexString(stateRoot), (state as BeaconStateAllForks) as BeaconStateAltair);
+          proofsServerApi.states.set(toHexString(stateRoot), (state as BeaconStateAllForks) as BeaconStateAltair);
         }
 
         // Emit a new head update with the custom state root
